All functions live
Got the compatibility calculator working; just added a separate file for the calculation to keep it cleaner. Added some jQuery to show the result - but I just read in the requirements that a modal is needed to show the result, so still work to do.  Made a simple home page. CSS tweaks.